python-selenium-tests/requirements.txt
pytest>=7.0
selenium>=4.20
python-dotenv>=1.0
# Install the local AccessFlow Python SDK wheel
../pythonSdk/accessflow_sdk-1.0.1-py3-none-any.whl
